+++ figma-API/icons-functions/icon-utility-files/icon-replacement-string.spec.js
@@ -1,0 +1,10 @@
+const { stringsToReplace } = require("./icon-replacement-string.js")
+
+describe("stringsToReplace", () => {
+    it("should not match social", () => {
+        expect(stringsToReplace.test("social")).toBeFalsy();
+    })
+    it("should not match social-incoming", () => {
+        expect(stringsToReplace.test("social-incoming")).toBeFalsy();
+    })
+})