+++ staging/src/components/Icons/iconInfo.tsx
@@ -2689,6 +2689,102 @@ export const icons = [
     category: "communication",
     animated: false,
   },
+  {
+    name: "audio-incoming",
+    bidirectional: false,
+    category: "communication",
+    animated: false
+  },
+  {
+    name: "audio-outgoing",
+    bidirectional: false,
+    category: "communication",
+    animated: false
+  },
+  {
+    name: "call-incoming",
+    bidirectional: false,
+    category: "communication",
+    animated: false
+  },
+  {
+    name: "call-outgoing",
+    bidirectional: false,
+    category: "communication",
+    animated: false
+  },
+  {
+    name: "chat-incoming",
+    bidirectional: false,
+    category: "communication",
+    animated: false
+  },
+  {
+    name: "chat-outgoing",
+    bidirectional: false,
+    category: "communication",
+    animated: false
+  },
+  {
+    name: "email-incoming",
+    bidirectional: false,
+    category: "communication",
+    animated: false
+  },
+  {
+    name: "email-outgoing",
+    bidirectional: false,
+    category: "communication",
+    animated: false
+  },
+  {
+    name: "messaging-incoming",
+    bidirectional: false,
+    category: "communication",
+    animated: false
+  },
+  {
+    name: "messaging-outgoing",
+    bidirectional: false,
+    category: "communication",
+    animated: false
+  },
+  {
+    name: "sms-incoming",
+    bidirectional: false,
+    category: "communication",
+    animated: false
+  },
+  {
+    name: "sms-outgoing",
+    bidirectional: false,
+    category: "communication",
+    animated: false
+  },
+  {
+    name: "social-incoming",
+    bidirectional: false,
+    category: "communication",
+    animated: false
+  },
+  {
+    name: "social-outgoing",
+    bidirectional: false,
+    category: "communication",
+    animated: false
+  },
+  {
+    name: "video-incoming",
+    bidirectional: false,
+    category: "communication",
+    animated: false
+  },
+  {
+    name: "video-outgoing",
+    bidirectional: false,
+    category: "communication",
+    animated: false
+  },  
   { name: "user", bidirectional: false, category: "account", animated: false },
   {
     name: "analytics-filled",